Rename JDBI 3 JUnit Jupiter extensions without Dropwizard in the name
The JDBI 3 Jupiter extensions have Dropwizard in their names but,
unlike the JDBI 2 extensions, they are independent of Dropwizard
and therefore can be renamed.

* DropwizardJdbi3DaoExtension -> Jdbi3DaoExtension
* DropwizardJdbi3Extension -> Jdbi3Extension
* DropwizardJdbi3Helpers -> Jdbi3Helpers

Fixes #117
